Move some gc code into headers
