+++ mycpp/gc_types.h
@@ -9,51 +9,7 @@
 
 namespace gc_heap {
 
-template <typename T>
-inline void InitSlabCell(Obj* obj) {
-  // log("SCANNED");
-  obj->heap_tag_ = Tag::Scanned;
-}
-
-template <>
-inline void InitSlabCell<int>(Obj* obj) {
-  // log("OPAQUE");
-  obj->heap_tag_ = Tag::Opaque;
-}
-
-// don't include items_[1]
-const int kSlabHeaderSize = sizeof(Obj);
-
-// Opaque slab, e.g. for List<int>
-template <typename T>
-class Slab : public Obj {
- public:
-  Slab(int obj_len) : Obj(0, 0, obj_len) {
-    InitSlabCell<T>(this);
-  }
-  T items_[1];  // variable length
-};
-
-template <typename T, int N>
-class GlobalSlab {
-  // A template type with the same layout as Str with length N-1 (which needs a
-  // buffer of size N).  For initializing global constant instances.
- public:
-  OBJ_HEADER()
-
-  T items_[N];
-
-  DISALLOW_COPY_AND_ASSIGN(GlobalSlab)
-};
-
-// Note: entries will be zero'd because the Heap is zero'd.
-template <typename T>
-inline Slab<T>* NewSlab(int len) {
-  int obj_len = RoundUp(kSlabHeaderSize + len * sizeof(T));
-  void* place = gHeap.Allocate(obj_len);
-  auto slab = new (place) Slab<T>(obj_len);  // placement new
-  return slab;
-}
+#include "mycpp/gc_slab.h"
 
 #ifdef MYLIB_LEAKY
   #define GLOBAL_STR(name, val) Str* name = StrFromC(val);
@@ -745,7 +701,6 @@ void Dict<K, V>::set(K key, V val) {
 // Do some extra calculation to avoid storing redundant lengths.
 inline int len(const gc_heap::Str* s) {
   assert(s->obj_len_ > gc_heap::kStrHeaderSize - 1);
-
   return s->obj_len_ - gc_heap::kStrHeaderSize - 1;
 }
